Add ability to update information about course and Top-10 students

+++ Course-Project/app/src/main/java/com/vladislavmyasnikov/courseproject/data/db/entities/StudentEntity.kt
@@ -2,11 +2,13 @@ package com.vladislavmyasnikov.courseproject.data.db.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.vladislavmyasnikov.courseproject.data.network.entities.MarkJson
+import com.vladislavmyasnikov.courseproject.domain.entities.Mark
 import com.vladislavmyasnikov.courseproject.domain.models.Identifiable
 
 @Entity(tableName = "students")
 data class StudentEntity(
         @PrimaryKey override val id: Int,
-         val name: String = "",
-         val mark: Double = 0.0
+        val name: String = "",
+        val mark: Float = 0.0f
 ) : Identifiable
